% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/llmixdist.r
\name{dll2}
\alias{dll2}
\title{A function that returns the probability density of a 2-log-logistic-mixture model}
\usage{
dll2(x, location1, scale1, location2, scale2, mixp, d)
}
\arguments{
\item{x}{A vector of the x values to evaluate the probability function}

\item{location1}{The location parameter of the first log-logistic distribution}

\item{scale1}{The scale parameter of the first log-logistic distribution}

\item{location2}{The location parameter of the second log-logistic distribution}

\item{scale2}{The scale parameter of the second log-logistic distribution}

\item{mixp}{The mixture probability of component 1}

\item{d}{The translational or shift parameter of the second log-logistic distribution}
}
\value{
A vector of the probability densities for the 2ll mixture
}
\description{
A function that returns the probability density of a 2-log-logistic-mixture model
}

