% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/llmixdist.r
\name{dllogis}
\alias{dllogis}
\title{The probability density function of the log-logistic distribution}
\usage{
dllogis(x, location, scale, d)
}
\arguments{
\item{x}{A vector of the x values to evaluate the probability function}

\item{location}{The location parameter of the log-logistic distribution}

\item{scale}{The scale parameter of the log-logistic distribution}

\item{d}{The translational or shift parameter, i.e., (x-d) follows the standard log-logistic distribution}
}
\description{
The probability density function of the log-logistic distribution
}

