% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/llmixdist.r
\name{pllogis}
\alias{pllogis}
\title{The log-logistic (cumulative) probility function}
\usage{
pllogis(q, location, scale, d, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{A vector of the x values to evaluate the probability function}

\item{location}{The location parameter of the log-logistic distribution}

\item{scale}{The scale parameter of the log-logistic distribution}

\item{d}{The translational or shift parameter, i.e., (q-d) follows the standard log-logistic distribution}

\item{lower.tail}{Whether to output the probability or survival function}

\item{log.p}{Whether to output the log of the probability (or survival) function}
}
\description{
The log-logistic (cumulative) probility function
}

