% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/llmixdist.r
\name{rllogis}
\alias{rllogis}
\title{A random log-logistic variable generator, by reversing the CDF}
\usage{
rllogis(n, location = 0, scale = 1)
}
\arguments{
\item{n}{The size of the return vector}

\item{location}{The location parameter, default = 0, which always returns 0}

\item{scale}{The scale parameter, defaulting to 1}
}
\value{
A vector of random values following the specified log-logistic distribution
}
\description{
A random log-logistic variable generator, by reversing the CDF
}

