% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/llmixdist.r
\name{sll2x}
\alias{sll2x}
\title{A function that returns the survival function for 2-log-logistic-mixture with eXpress saccade component}
\usage{
sll2x(x, location1, scale1, location2, scale2, mixp, d, px)
}
\arguments{
\item{x}{A vector of the x values to evaluate the probability function}

\item{location1}{The location parameter of the first log-logistic distribution}

\item{scale1}{The scale parameter of the first log-logistic distribution}

\item{location2}{The location parameter of the second log-logistic distribution}

\item{scale2}{The scale parameter of the second log-logistic distribution}

\item{mixp}{The mixture probability of component 1; (1-mixp) is the proportion of component 2}

\item{d}{The translational or shift parameter of the second log-logistic distribution}

\item{px}{The mixing probability of the eXpress component; (1-px) are the proportion of RTs for the 2ll mixture}

\item{lx}{The location parameter for the eXpress component, which can be set by the global lx.default}

\item{sx}{The scale parameter for the eXpress component, which can be set by the global sx.default}
}
\value{
A vector of the probability densities for the ll2x mixture
}
\description{
The "eXpress" component often appears in the data may be a result of noise or corrective
  responses that have little to do with the present task.
Note the mixture probabilities are set up in the following way:
  px + (1-px)(mixp + 1-mixp) === 100%
}

